+++ examples/live/binance_futures_testnet_ema_cross_with_trailing_stop.py
@@ -87,7 +87,7 @@ strat_config = EMACrossTrailingStopConfig(
     atr_period=20,
     trailing_atr_multiple=3.0,
     trailing_offset_type="BASIS_POINTS",
-    trigger_type="LAST",
+    trigger_type="LAST_TRADE",
     trade_size=Decimal("0.010"),
 )
 # Instantiate your strategy