+++ nautilus_trader/core/rust/model.pyx
@@ -35,3 +35,4 @@ from nautilus_trader.core.rust.model cimport PositionSide  # type: ignore
 from nautilus_trader.core.rust.model cimport PriceType  # type: ignore
 from nautilus_trader.core.rust.model cimport TimeInForce  # type: ignore
 from nautilus_trader.core.rust.model cimport TrailingOffsetType  # type: ignore
+from nautilus_trader.core.rust.model cimport TriggerType  # type: ignore