+++ tests/unit_tests/backtest/test_backtest_exchange_bracket_if_touched_entries.py
@@ -163,7 +163,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "emulation_trigger, " "order_side, " "sl_trigger_price, " "tp_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3050.00,
                 3150.00,
@@ -175,7 +175,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3150.00,
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3150.00,
                 3050.00,
@@ -234,7 +234,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "next_tick_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3090.00,  # entry_trigger_price
                 3050.00,  # sl_trigger_price
@@ -250,7 +250,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3090.00,  # next_tick_price (hits trigger)
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3110.00,  # entry_trigger_price,
                 3150.00,  # sl_trigger_price
@@ -333,7 +333,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "next_tick_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3090.00,  # entry_trigger_price
                 3089.00,  # entry_price
@@ -351,7 +351,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3090.00,  # next_tick_price (hits trigger)
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3110.00,  # entry_trigger_price,
                 3111.00,  # entry_price
@@ -449,7 +449,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "next_tick_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3099.00,  # entry_trigger_price
                 3095.00,  # entry_price
@@ -467,7 +467,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3094.00,  # next_tick_price (moves through limit price)
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3101.00,  # entry_trigger_price
                 3105.00,  # entry price
@@ -554,7 +554,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "next_tick_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3099.00,  # entry_trigger_price
                 3100.00,  # entry_price
@@ -572,7 +572,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3098.00,  # next_tick_price (moves through trigger price)
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3101.00,  # entry_trigger_price
                 3099.00,  # entry price
@@ -659,7 +659,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "next_tick_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3100.00,  # entry_trigger_price (triggers immediately)
                 3095.00,  # entry_price
@@ -677,7 +677,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3094.00,  # next_tick_price (moves through limit price)
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3100.00,  # entry_trigger_price (triggers immediately)
                 3105.00,  # entry price
@@ -763,7 +763,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
         "tp_price",
         [
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.BUY,
                 3102.00,  # entry_trigger_price (triggers immediately)
                 3103.00,  # entry_price
@@ -779,7 +779,7 @@ class TestSimulatedExchangeEmulatedContingencyOrders:
                 3150.00,  # tp_price
             ],
             [
-                TriggerType.NONE,
+                TriggerType.NO_TRIGGER,
                 OrderSide.SELL,
                 3098.00,  # entry_trigger_price (triggers immediately)
                 3097.00,  # entry_price